# Improve TruncateRowNumber