@joe-elliott 2023-08-15
Rerunning benchmarks and I'm seeing mixed results:

```
name                                                old time/op    new time/op    delta
BackendBlockTraceQL/spanAttNameNoMatch-8              7.85ms ± 3%    7.62ms ± 0%   -2.88%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttValNoMatch-8               13.8ms ± 1%    13.4ms ± 1%   -3.29%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttValMatch-8                  422ms ± 1%     450ms ± 3%   +6.63%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8         8.25ms ± 2%    7.60ms ± 2%   -7.84%  (p=0.008 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicMatch-8            368ms ± 3%     376ms ± 3%     ~     (p=0.095 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicRegexNoMatch-8    8.83ms ± 1%    8.21ms ± 1%   -7.03%  (p=0.029 n=4+4)
BackendBlockTraceQL/spanAttIntrinsicRegexMatch-8       593ms ±12%     579ms ± 3%     ~     (p=0.841 n=5+5)
BackendBlockTraceQL/resourceAttNameNoMatch-8          6.72ms ± 2%    6.43ms ± 1%   -4.30%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttValNoMatch-8           29.4ms ±11%    27.4ms ± 1%   -6.82%  (p=0.008 n=5+5)
BackendBlockTraceQL/resourceAttValMatch-8              587ms ± 5%     572ms ± 3%     ~     (p=0.548 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicNoMatch-8     6.49ms ± 0%    6.47ms ± 2%     ~     (p=0.421 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8        198ms ± 2%     201ms ± 2%     ~     (p=0.095 n=5+5)
BackendBlockTraceQL/mixedNameNoMatch-8                 3.79s ± 2%     3.62s ± 3%   -4.54%  (p=0.016 n=5+5)
BackendBlockTraceQL/mixedValNoMatch-8                  3.02s ± 2%     2.99s ± 2%     ~     (p=0.690 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchAnd-8           6.64ms ± 1%    7.05ms ±10%     ~     (p=0.690 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr-8             3.14s ± 1%     3.15s ± 1%     ~     (p=0.841 n=5+5)
BackendBlockTraceQL/mixedValBothMatch-8               6.48ms ± 1%    7.25ms ± 1%  +11.92%  (p=0.008 n=5+5)
```

I think it might make sense to let this one go.